fix(build): restore Minecraft 26.1.2 compilation
- migrate renamed and relocated Minecraft 26.1.2 types using local game sources
- preserve potion, book, container, and vehicle behavior across API changes
- add the JUnit Platform launcher to subproject test runtime classpaths
- wire generated message keys into the sources JAR task graph

+++ cellulosesz-modules/cellulosesz-module-sign/src/main/java/top/likoslupus/cellulosesz/modules/sign/handler/RepairSignHandler.java
@@ -40,7 +40,7 @@ public final class RepairSignHandler implements SynchronousSignHandler {
                 ? RepairScope.ALL
                 : RepairScope.HAND;
         var result = items.repair(context.player(), scope);
-        var count = result.value().orElse(0);
+        var count = (int) result.value().orElse(0);
 
         return result.successful() && count > 0
                 ?
